backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Ali Elganzory <alielganzory@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Flask = "^2.2.3"
Flask-SQLAlchemy = "^3.0.3"
celery = {extras = ["redis"], version = "^5.2.7"}
blinker = "^1.6.2"
autopep8 = "^2.0.2"
numpy = "^1.24.2"
ultralytics = "^8.0.81"
flask-cors = "^3.0.10"

[tool.poetry.group.dev.dependencies]
watchdog = "^3.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
